Working Dockerfile
